current age for this year

+++ Simple_ruby/exercise_9.rb
@@ -1,9 +1,12 @@
-# ruby simple_ruby/exercise_9.rb Write a Ruby program that takes a year as input and checks whether it is a leap year. Print the result.
+# ruby simple_ruby/exercise_9.rb Write a Ruby program
+# that takes a year as input and checks whether it is a leap year.
+# Print the result.
 
-  year = gets.chomp.to_f
-if
-year % 4 ==0
+  year = gets.chomp.to_i
+
+if year % 4 == 0
   print "leap year"
 else
   print "not leap year"
 end
+
